Exception handling and KeyMap class additions
Added exception handling to the evaluatepedals method when I noticed
exceptions during use.

Started setting up the KeyMap class for testing.

Added a check in the EvaluatePedals method to see if the pedal value is
at or greater than the threshold and to that the key is not currently
down. This prevents the program from spamming the key down event.