KeyMap additions and started implementing customizable serial ports
Added FilterKey() method to the MainWindow class. It is used by the
PreviewKeyDown() events to filter out invalid keystrokes.

Implemented configurable keys for brake and clutch.

Started implementing configurable COM port. This is currently broken.

Removed while loop in MainWindow constructor as it was not needed.

Need to finish configurable COM port and saving of settings.